// Jobs.Frontend/src/app/core/api/model/createJobPostingRequest.ts
/**
 * Jobs API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JobCategory } from './jobCategory';


/**
 * Request DTO for creating a job posting.
 */
export interface CreateJobPostingRequest { 
    title: string;
    description?: string | null;
    category?: JobCategory;
    minHoursPerWeek?: number | null;
    maxHoursPerWeek?: number | null;
    allowsRemoteWork?: boolean | null;
    contactEmail?: string | null;
    jobUrl?: string | null;
    companyId: string;
}
export namespace CreateJobPostingRequest {
}